refactor: Split string functions module into focused submodules
Reorganized the monolithic string.rs (563 lines) into 8 smaller, focused modules:

**Modules created:**
- case.rs (47 lines) - UPPER, LOWER case conversion
- concat.rs (36 lines) - CONCAT string concatenation
- length.rs (90 lines) - LENGTH, CHAR_LENGTH, OCTET_LENGTH
- search.rs (119 lines) - POSITION, INSTR, LOCATE string searching
- substring.rs (162 lines) - SUBSTRING, LEFT, RIGHT extraction
- transform.rs (52 lines) - REPLACE, REVERSE transformations
- trim.rs (87 lines) - TRIM operations with advanced options
- mod.rs (33 lines) - Module organization and re-exports

**Benefits:**
- Improved code organization by functional category
- Easier navigation and maintenance
- Better separation of concerns
- Most modules under 150 lines (7 of 8 meet criteria)
- All string functions work identically (tests pass)
- No performance degradation
- Maintains SQL:1999 standard compliance documentation
